+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 ### Features
 - feat: initial release of a11yer
 
+### Fixes
+- fix: CI peer deps + Playwright cache
+
+
+## Unreleased (2026-03-30)
+
+### Features
+- feat: initial release of a11yer
+
 
 ## 0.1.0 (2026-03-30)